refactor: .svg 이미지 avif 로 변경
